pyproject.toml
[tool.poetry]
name = "gestor-tareas-web"
version = "1.0.0"
description = "Gestor de Tareas Web - Task Manager Application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.0"
flask-jwt-extended = "^2.10.1"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.5"
flask-wtf = "*"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"
werkzeug = "^3.1.3"
email-validator = "^2.0.0"
alembic = "^1.16.4"
bcrypt = "^4.3.0"
blinker = "^1.9.0"
click = "^8.2.2"
greenlet = "^3.2.3"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.6"
markupsafe = "^3.0.2"
sqlalchemy = "^2.0.42"
typing-extensions = "^4.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"